requirements.txt
pandas 
gdown
setuptools
numpy
pyYAML
tqdm
joblib
types-PyYAML
scipy
Flask
Flask-Cors
scikit-learn
requests
python-dotenv
mysql-connector-
boto3
sagemaker
-e .
